+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated `cargo contract new` template dependencies to ink! `version = "3"` - [#569](https://github.com/paritytech/cargo-contract/pull/569)
+- Improved documentation on how to invoke `cargo contract decode` - [#572](https://github.com/paritytech/cargo-contract/pull/572)
 
 ### Fixed
 - Make constructor selector look for exact function name - [#562](https://github.com/paritytech/cargo-contract/pull/562) (thanks [@forgetso](https://github.com/forgetso)!)
+- Fix dirty directory issue when crate installation had been interrupted - [#571](https://github.com/paritytech/cargo-contract/pull/571)
 
 ## [1.3.0] - 2022-05-09
 
@@ -58,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Interact with contracts: upload, instantiate and call commands
 
 We added commands to upload, instantiate and call contracts!
-This allows interacting with contracts on live chains with a compatible 
+This allows interacting with contracts on live chains with a compatible
 [`pallet-contracts`](https://github.com/paritytech/substrate/tree/master/frame/contracts).
 
 For command-line examples on how to use these commands see [#79](https://github.com/paritytech/cargo-contract/pull/79).